requirements.txt
docx2pdf==0.1.8
numpy==2.1.3
openpyxl==3.1.5
pandas==2.2.3
python_pptx==1.0.2
comtypes==1.4.8